cliq/requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
google-api-python-client
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
mysqlclient==1.4.2.post1
Pillow==6.0.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
